GoLang: Add simple writer that mimics the C# bench
It is only a naive implementation of a Go benchmark following the same
pattern as the C# benchmark meant to have something to compare the C#
to. It probably can be improved a lot in terms of both performance and
style, given the fact that I am not proficient at all in Go.